+++ bin/actions/0220022-106-cr.sh
@@ -1,0 +1,14 @@
+#!/bin/bash
+
+#: Right-CTRL + ◢ + Right - Chromium/Firefox: Focus next tab, than reload it
+
+focusApplication=\
+$(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
+
+if [[ "$focusApplication" =~ "chromium-browse" ]] \
+|| [[ "$focusApplication" =~ "firefox" ]]; then
+  xdotool key ctrl+Tab
+  xdotool sleep 0.1
+  xdotool key ctrl+F5
+fi
+